# docs/swagger.yaml
basePath: /
definitions:
...
  github_com_proarash_ecommerce-server_internal_types.ApiResponse:
    properties:
      data: {}
      message:
        type: string
    type: object
  github_com_proarash_ecommerce-server_internal_types.ErrorResponse:
    properties:
      error:
        type: string
    type: object
  github_com_proarash_ecommerce-server_internal_types.MessageResponse:
    properties:
      message:
        type: string
    type: object
  github_com_proarash_ecommerce-server_internal_types.PaginatedResponse:
    properties:
      items: {}
      limit:
        type: integer
      page:
        type: integer
      total:
        type: integer
    type: object
  github_com_proarash_ecommerce-server_internal_user.User:
    properties:
      address:
        type: string
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      id:
        type: integer
      latitude:
        type: number
      longitude:
        type: number
      mobile:
        type: string
      name:
        type: string
      status:
        type: boolean
      telegram_chat_id:
        type: integer
      updatedAt:
        type: string
      wallet:
        $ref: '#/definitions/github_com_proarash_ecommerce-server_internal_wallet.Wallet'
    type: object
  github_com_proarash_ecommerce-server_internal_wallet.Wallet:
    properties:
      amount:
        type: number
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      id:
        type: integer
      locked:
        type: boolean
      owner_id:
        type: integer
      owner_type:
        enum:
        - user
        - staff
        type: string
      updatedAt:
        type: string
    type: object
  gorm.DeletedAt:
    properties:
      time:
        type: string
      valid:
        description: Valid is true if Time is not NULL
        type: boolean
    type: object
...
  internal_discount.CheckDto:
    properties:
      amount:
        example: 200000
        type: number
      code:
        example: SUMMER30
        maxLength: 32
        type: string
    required:
    - amount
    - code
    type: object
  internal_discount.CheckResponse:
    properties:
      amount:
        type: number
      code:
        type: string
      discount_amount:
        type: number
      payable_amount:
        type: number
      type:
        enum:
        - percent
        - value
        type: string
    type: object
  internal_discount.CreateDiscountDto:
    properties:
      code:
        example: SUMMER30
        maxLength: 32
        minLength: 4
        type: string
      max_price:
        example: 50000
        type: number
      owner_mobile:
        example: "09121112233"
        type: string
      status:
        type: boolean
      type:
        enum:
        - percent
        - value
        type: string
      use_count:
        example: 100
        minimum: 1
        type: integer
      value:
        example: 30
        type: number
    required:
    - type
    - use_count
    - value
    type: object
  internal_discount.Discount:
    properties:
      code:
        example: SUMMER30
        type: string
      created_by_id:
        type: integer
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      id:
        type: integer
      max_price:
        type: number
      owner:
        $ref: '#/definitions/github_com_proarash_ecommerce-server_internal_user.User'
      owner_id:
        type: integer
      status:
        type: boolean
      type:
        enum:
        - percent
        - value
        type: string
      updatedAt:
        type: string
      use_count:
        type: integer
      used_count:
        type: integer
      value:
        type: number
    type: object
  internal_discount.UpdateDiscountDto:
    properties:
      max_price:
        minimum: 0
        type: number
      owner_mobile:
        type: string
      status:
        type: boolean
      type:
        enum:
        - percent
        - value
        type: string
      use_count:
        minimum: 1
        type: integer
      value:
        type: number
    type: object
...
host: localhost:4000
info:
  contact: {}
  description: 'E-commerce backend: catalog, cart, orders, Zibal payments, inventory,
    CMS, support chat and notifications. All JSON responses are wrapped in {"message":
    string, "data": any}.'
  title: E-Commerce Server API
  version: "1.0"
paths:
...
  /discounts:
    get:
      description: Available to admin, accountant, marketer and support (read-only)
      parameters:
      - description: Session to authenticate with; selects the staff or customer auth
          cookie
        enum:
        - staff
        - customer
        in: query
        name: user_type
        type: string
      - description: Code contains
        in: query
        name: code
        type: string
      - description: Type
        enum:
        - percent
        - value
        in: query
        name: type
        type: string
      - description: Status
        in: query
        name: status
        type: boolean
      - description: Owner user ID
        in: query
        name: owner_id
        type: integer
      - default: 1
        description: Page
        in: query
        name: page
        type: integer
      - default: 20
        description: Limit
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_proarash_ecommerce-server_internal_types.ApiResponse'
            - properties:
                data:
                  allOf:
                  - $ref: '#/definitions/github_com_proarash_ecommerce-server_internal_types.PaginatedResponse'
                  - properties:
                      items:
                        items:
                          $ref: '#/definitions/internal_discount.Discount'
                        type: array
                    type: object
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/github_com_proarash_ecommerce-server_internal_types.ApiResponse'
            - properties:
                data:
                  $ref: '#/definitions/github_com_proarash_ecommerce-server_internal_types.ErrorResponse'
              type: object
        "403":
          description: Forbidden
          schema:
            allOf:
            - $ref: '#/definitions/github_com_proarash_ecommerce-server_internal_types.ApiResponse'
            - properties:
                data:
                  $ref: '#/definitions/github_com_proarash_ecommerce-server_internal_types.ErrorResponse'
              type: object
      security:
      - BearerAuth: []
      summary: List discount codes
      tags:
      - Discounts
    post:
      consumes:
      - application/json
      description: Creates a discount code (admin, accountant, marketer). Omit code
        to auto-generate an 8 character uppercase code. max_price caps the discount
        and is only allowed for percent type (e.g. 30% up to 50000). When owner_mobile
        is set only the customer with that mobile can use the code.
      parameters:
      - description: Session to authenticate with; selects the staff or customer auth
          cookie
        enum:
        - staff
        - customer
        in: query
        name: user_type
        type: string
      - description: Discount
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_discount.CreateDiscountDto'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/github_com_proarash_ecommerce-server_internal_types.ApiResponse'
            - properties:
                data:
                  $ref: '#/definitions/internal_discount.Discount'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/github_com_proarash_ecommerce-server_internal_types.ApiResponse'
            - properties:
                data:
                  $ref: '#/definitions/github_com_proarash_ecommerce-server_internal_types.ErrorResponse'
              type: object
        "403":
          description: Forbidden
          schema:
            allOf:
            - $ref: '#/definitions/github_com_proarash_ecommerce-server_internal_types.ApiResponse'
            - properties:
                data:
                  $ref: '#/definitions/github_com_proarash_ecommerce-server_internal_types.ErrorResponse'
              type: object
        "409":
          description: Conflict
          schema:
            allOf:
            - $ref: '#/definitions/github_com_proarash_ecommerce-server_internal_types.ApiResponse'
            - properties:
                data:
                  $ref: '#/definitions/github_com_proarash_ecommerce-server_internal_types.ErrorResponse'
              type: object
      security:
      - BearerAuth: []
      summary: Generate discount code
      tags:
      - Discounts
  /discounts/{id}:
    delete:
      parameters:
      - description: Session to authenticate with; selects the staff or customer auth
          cookie
        enum:
        - staff
        - customer
        in: query
        name: user_type
        type: string
      - description: Discount ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_proarash_ecommerce-server_internal_types.ApiResponse'
            - properties:
                data:
                  $ref: '#/definitions/github_com_proarash_ecommerce-server_internal_types.MessageResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/github_com_proarash_ecommerce-server_internal_types.ApiResponse'
            - properties:
                data:
                  $ref: '#/definitions/github_com_proarash_ecommerce-server_internal_types.ErrorResponse'
              type: object
        "403":
          description: Forbidden
          schema:
            allOf:
            - $ref: '#/definitions/github_com_proarash_ecommerce-server_internal_types.ApiResponse'
            - properties:
                data:
                  $ref: '#/definitions/github_com_proarash_ecommerce-server_internal_types.ErrorResponse'
              type: object
        "404":
          description: Not Found
          schema:
            allOf:
            - $ref: '#/definitions/github_com_proarash_ecommerce-server_internal_types.ApiResponse'
            - properties:
                data:
                  $ref: '#/definitions/github_com_proarash_ecommerce-server_internal_types.ErrorResponse'
              type: object
      security:
      - BearerAuth: []
      summary: Delete discount code
      tags:
      - Discounts
    get:
      description: Available to admin, accountant, marketer and support (read-only)
      parameters:
      - description: Session to authenticate with; selects the staff or customer auth
          cookie
        enum:
        - staff
        - customer
        in: query
        name: user_type
        type: string
      - description: Discount ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_proarash_ecommerce-server_internal_types.ApiResponse'
            - properties:
                data:
                  $ref: '#/definitions/internal_discount.Discount'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/github_com_proarash_ecommerce-server_internal_types.ApiResponse'
            - properties:
                data:
                  $ref: '#/definitions/github_com_proarash_ecommerce-server_internal_types.ErrorResponse'
              type: object
        "403":
          description: Forbidden
          schema:
            allOf:
            - $ref: '#/definitions/github_com_proarash_ecommerce-server_internal_types.ApiResponse'
            - properties:
                data:
                  $ref: '#/definitions/github_com_proarash_ecommerce-server_internal_types.ErrorResponse'
              type: object
        "404":
          description: Not Found
          schema:
            allOf:
            - $ref: '#/definitions/github_com_proarash_ecommerce-server_internal_types.ApiResponse'
            - properties:
                data:
                  $ref: '#/definitions/github_com_proarash_ecommerce-server_internal_types.ErrorResponse'
              type: object
      security:
      - BearerAuth: []
      summary: Get discount code
      tags:
      - Discounts
    patch:
      consumes:
      - application/json
      description: Updates a discount (admin, accountant, marketer). Send max_price
        0 to remove the cap and owner_mobile "" to make the code public.
      parameters:
      - description: Session to authenticate with; selects the staff or customer auth
          cookie
        enum:
        - staff
        - customer
        in: query
        name: user_type
        type: string
      - description: Discount ID
        in: path
        name: id
        required: true
        type: integer
      - description: Fields to update
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_discount.UpdateDiscountDto'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_proarash_ecommerce-server_internal_types.ApiResponse'
            - properties:
                data:
                  $ref: '#/definitions/internal_discount.Discount'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/github_com_proarash_ecommerce-server_internal_types.ApiResponse'
            - properties:
                data:
                  $ref: '#/definitions/github_com_proarash_ecommerce-server_internal_types.ErrorResponse'
              type: object
        "403":
          description: Forbidden
          schema:
            allOf:
            - $ref: '#/definitions/github_com_proarash_ecommerce-server_internal_types.ApiResponse'
            - properties:
                data:
                  $ref: '#/definitions/github_com_proarash_ecommerce-server_internal_types.ErrorResponse'
              type: object
        "404":
          description: Not Found
          schema:
            allOf:
            - $ref: '#/definitions/github_com_proarash_ecommerce-server_internal_types.ApiResponse'
            - properties:
                data:
                  $ref: '#/definitions/github_com_proarash_ecommerce-server_internal_types.ErrorResponse'
              type: object
      security:
      - BearerAuth: []
      summary: Update discount code
      tags:
      - Discounts
...
  /user/discounts/check:
    post:
      consumes:
      - application/json
      description: Validates a discount code for the current customer against an amount
        and returns the discount and payable amount without redeeming it
      parameters:
      - description: Session to authenticate with; selects the staff or customer auth
          cookie
        enum:
        - staff
        - customer
        in: query
        name: user_type
        type: string
      - description: Code and amount
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_discount.CheckDto'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_proarash_ecommerce-server_internal_types.ApiResponse'
            - properties:
                data:
                  $ref: '#/definitions/internal_discount.CheckResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/github_com_proarash_ecommerce-server_internal_types.ApiResponse'
            - properties:
                data:
                  $ref: '#/definitions/github_com_proarash_ecommerce-server_internal_types.ErrorResponse'
              type: object
        "403":
          description: Forbidden
          schema:
            allOf:
            - $ref: '#/definitions/github_com_proarash_ecommerce-server_internal_types.ApiResponse'
            - properties:
                data:
                  $ref: '#/definitions/github_com_proarash_ecommerce-server_internal_types.ErrorResponse'
              type: object
      security:
      - BearerAuth: []
      summary: Check discount code
      tags:
      - User
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and the JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"